fix(codegen): stack limitation should be 1024 (#295)

+++ codegen/src/visitor/call.rs
@@ -89,7 +89,7 @@ impl Function {
 
         // Adjust the stack pointer for the results.
         self.masm._jumpdest()?;
-        self.masm.increment_sp(*results as u8)?;
+        self.masm.increment_sp(*results as u16)?;
         Ok(())
     }
